able to fetch all users in new all-users file
